requirements.txt
attrs
requests
fastapi
uvicorn
python-dotenv
pydantic-settings
langchain
openai
pdfplumber
python-docx
python-multipart
beautifulsoup4
aiohttp
sqlalchemy
pymysql
redis
lxml
pytest
pytest-mock